+++ internal/entity/fault.go
@@ -7,7 +7,7 @@ import (
 
 type Fault struct {
 	ErrType   string    `json:"type"`
-	UserID    uint16    `json:"userID"`
+	UserID    uint32    `json:"userID"`
 	Date      time.Time `json:"date"`
 	UserAgent string    `json:"userAgent"`
 }
