+++ internal/entity/love.go
@@ -7,8 +7,8 @@ import (
 
 type Love struct {
 	BoardID  uint16    `json:"boardID"`
-	Timecode uint16    `json:"timecode"`
-	UserID   uint16    `json:"userID"`
+	Timecode uint32    `json:"timecode"`
+	UserID   uint32    `json:"userID"`
 	Date     time.Time `json:"date"`
 }
